+++ tests/cases/tabs/test_prepare.py
@@ -17,9 +17,9 @@ os.environ["PYTEST_QT_API"] = "pyqt5"
 # move to the fast
 
 
-@pytest.mark.dependency(depends=["tabs_bootstrap_ui"])
+@pytest.mark.dependency(depends=["tabs_bootstrap_ui"], scope="session")
 class TestREvoDesignPlugin_TabPrepare:
-    @pytest.mark.dependency(name="tabs_bootstrap_prepare", depends=["tabs_bootstrap_ui"])
+    @pytest.mark.dependency(name="tabs_bootstrap_prepare", depends=["tabs_bootstrap_ui"], scope="session")
     def test_load_molecule(self, test_worker: TestWorker):
         test_worker.test_id = test_worker.method_name()
         test_worker.load_session_and_check(from_rcsb=True)
@@ -32,6 +32,7 @@ class TestREvoDesignPlugin_TabPrepare:
         test_worker.save_pymol_png(basename=test_worker.test_id)
         test_worker.save_new_experiment()
 
+    @pytest.mark.dependency(name="tabs_prepare_pocket_session", depends=["tabs_bootstrap_prepare"], scope="session")
     def test_pocket(self, test_worker: TestWorker):
         test_worker.test_id = test_worker.method_name()
         test_worker.load_session_and_check(from_rcsb=True)